+++ src/helpers.js
@@ -546,17 +546,27 @@ export function solveUpdate(affectedPaths, comparedPaths) {
 /**
  * Non-mutative version of the splice array method.
  *
- * @param {array}    array        - The array to splice.
- * @param {integer}  startIndex   - The start index.
- * @param {integer}  nb           - Number of elements to remove.
- * @param {...mixed} elements     - Elements to append after splicing.
- * @return {array}                - The spliced array.
+ * @param  {array}    array        - The array to splice.
+ * @param  {integer}  startIndex   - The start index.
+ * @param  {integer}  nb           - Number of elements to remove.
+ * @param  {...mixed} elements     - Elements to append after splicing.
+ * @return {array}                 - The spliced array.
  */
 export function splice(array, startIndex, nb, ...elements) {
+  nb = Math.max(0, nb);
+
+  // Positive index
+  if (startIndex >= 0)
+    return array
+      .slice(0, startIndex)
+      .concat(elements)
+      .concat(array.slice(startIndex + nb));
+
+  // Negative index
   return array
-    .slice(0, startIndex)
+    .slice(0, array.length + startIndex)
     .concat(elements)
-    .concat(array.slice(startIndex + Math.max(0, nb)));
+    .concat(array.slice(array.length + startIndex + nb));
 }
 
 /**