+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adding the `tree/cursor.clone` and `tree/cursor.deepClone` methods.
 * Fixing an issue where the `tree.commit` method would fire a useless update.
 * Fixing an issue related to updates and dynamic paths.
+* Fixing the `tree/cursor.splice` to correctly handle negative indexes.
 
 ## v2.2.1